refactor: OPTIC-429: Remove unused FF ff_front_dev_1470_dm_pagination_010422_short (#305)
