+++ src/components/MainView/GridView/GridView.styl
@@ -30,9 +30,6 @@
       border-radius 2px
       background none
 
-      img
-        object-fit contain
-
       &::after
         top 0
         left 0
