+++ src/utils/feature-flags.js
@@ -1,9 +1,6 @@
 // Outliner + Details
 export const FF_DEV_1170 = "ff_front_1170_outliner_030222_short";
 
-// Switch to page navigation
-export const FF_DEV_1470 = "ff_front_dev_1470_dm_pagination_010422_short";
-
 /**
  * Support for notification links in the Label Steam and the Review Stream.
  * @link https://app.launchdarkly.com/default/branch/features/feat_front_dev_1752_notification_links_in_label_and_review_streams
